requirements.txt
google-genai>=1.0.0
PyGithub>=2.1.1
GitPython>=3.1.40
PyYAML>=6.0.1
